requirements.txt
numpy
pandas
Pillow
plotly
